Fix a bug in the rounding logic
When we firstly found a 2-digits sequence that is not "99" during rounding up, what we need to do is to add 1 to it. If the second digit of those 2-digits is not '9', then we can just add 1 to it, but if it's '9', we need to replace it by '0' and add 1 to the first digit.